Make Identifiers generically, lazily composable
Previously, HttpConfig had logic for composing Identifiers. This is moved into
the RoutingFactory so it may be used across protocols.

Furthermore, identifier composition was eager, identifying each request through
all identifiers and then checking the results in order.  Now, identifiers are
checked serially. This change is made under the assumption that Identifiers
tend to be fast but incur allocation.